+++ Dissasemble/AJMP.h
@@ -22,11 +22,11 @@ public:
         bytes.push_back(*istr);
         istr++;
 
-        stream << "AJMP    " << std::setfill('0') << std::setw(4) << std::hex << (uint16_t) addr;
+        stream << "AJMP    " << std::setfill('0') << std::setw(4) << std::hex << (uint32_t) addr;
         toString = stream.str();
     }
 
-    uint16_t addr;
+    uint32_t addr;
 
 };
