+++ Dissasemble/CJNE.h
@@ -18,13 +18,13 @@ public:
         addr = 3 + pos + *(std::vector<int8_t>::iterator &) istr;
         bytes.push_back(*istr);
         istr++;
-        stream << "  " << std::setfill('0') << std::setw(4) << std::hex << (uint16_t) addr;
+        stream << "  " << std::setfill('0') << std::setw(4) << std::hex << (uint32_t) addr;
         toString = stream.str();
         bytes.push_back(*istr);
         istr++;
     }
 
-    uint16_t addr;
+    uint32_t addr;
 };
 
 #endif //CC2530DIS_CJNE_H
