+++ Dissasemble/JCOND.h
@@ -48,10 +48,10 @@ public:
         }
         bytes.push_back(*istr);
         istr++;
-        stream << std::setfill('0') << std::setw(4) << std::hex << (uint16_t) addr;
+        stream << std::setfill('0') << std::setw(4) << std::hex << (uint32_t) addr;
         toString = stream.str();
     }
-    uint16_t  addr;
+    uint32_t  addr;
     uint8_t bit;
 };
